Move AssetCommandSetView from assets to mission
The command-set endpoint is a mission feature; it uses mission_is_member,
AssetCommand, MissionAsset, and AssetCommandForm. Moving it to
mission/views.py and mission/urls.py (keeping name='asset_command_set')
removes mission_is_member, MissionAsset, AssetCommandForm, and Point
from assets/views.py imports.
