+++ mission/urls.py
@@ -25,4 +25,5 @@ urlpatterns = [
     re_path(r'^mission/list/$', views.mission_list_data, name='mission_list_data'),
     re_path(r'^mission/asset/status/values/$', views.MissionAssetStatusValuesView.as_view()),
     re_path(r'^$', views.mission_list, name='mission_list'),
+    re_path(r'^mission/(?P<mission_id>\d+)/assets/command/set/$', views.AssetCommandSetView.as_view(), name='asset_command_set'),
 ]
